backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
cryptography>=42.0.0
Pillow>=10.2.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
supabase>=2.0.0
